tests on parser

+++ server/parser/scene_creation.c
@@ -43,7 +43,7 @@ static t_scene		extend_scene(int fd, t_scene scene, char **line, t_env *env)
 		scene.theme = color(ft_atoi(line[1]), ft_atoi(line[2]),
 						ft_atoi(line[3]), 0);
 	else if (ft_strequ(line[0], "power") && line_len(line) == 2)
-		scene.power = fmin(fmax(ft_atoi(line[1]) / 100.0, 0), 1);
+		scene.power = fmin(fmax((float)ft_atoi(line[1]) / 100.0, 0), 1);
 	if (line_len(line) != 2 || !(ft_strequ(line[1], "{")))
 		return (scene);
 	if (ft_strequ(line[0], "light"))
